requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1.2
django-allauth==0.51.0
django-filter==22.1
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
Markdown==3.4.1
oauthlib==3.2.2
psycopg2==2.9.4
pycparser==2.21
PyJWT==2.5.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
types-cryptography==3.3.23.1
urllib3==1.26.12
whitenoise==6.2.0
zipp==3.9.0
